requirements.txt
django==3.2.10
google-api-python-client
psycopg2
apscheduler
djangorestframework